change value of excel report

+++ database/migrations/2023_01_13_222635_alter_total_qty_to_waste_logs_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('waste_logs', function (Blueprint $table) {
-            $table->float('total_qty')->default(0);
+            $table->double('total_qty')->default(0);
         });
     }
